+++ packages/dumb/src/index.ts
@@ -77,6 +77,7 @@ import { buildDevice } from './utils/device/device.utils'
 import { SafetyBasalItem } from './models/safety-basal-item.model'
 import { isSafetyBasalAvailable, getSafetyBasalItems } from './utils/safety-basal-profile/safety-basal-profile.util'
 import { NightModeTooltip } from './components/tooltips/night-mode-tooltip/night-mode-tooltip'
+import { SimpleValue } from './components/stats/common/simple-value'
 
 export {
   formatDate,
@@ -126,5 +127,6 @@ export {
   WarmUpTooltip,
   SafetyBasalItem,
   isSafetyBasalAvailable,
-  getSafetyBasalItems
+  getSafetyBasalItems,
+  SimpleValue
 }